refactor: restructure files to common src pattern
